replace mocha with jest
