+++ .babelrc
@@ -14,7 +14,10 @@
   ],
   "env": {
     "test": {
-      "presets": ["@babel/preset-env", "@babel/preset-react"],
+      "presets": [
+        "@babel/preset-env",
+        "@babel/preset-react"
+      ],
       "plugins": [
         "transform-class-properties",
         "transform-es2015-spread",
