+++ .eslintrc
@@ -21,9 +21,16 @@
         "comma-dangle": 0,
         "class-methods-use-this": 0,
         "func-names": 0,
-        "indent": [2, 4],
+        "import/no-extraneous-dependencies": 0,
+        "indent": [
+            2,
+            4
+        ],
         "jsx-a11y/no-static-element-interactions": 0,
-        "max-len": [2, 120],
+        "max-len": [
+            2,
+            120
+        ],
         "no-tabs": 0,
         "react/jsx-indent": 0,
         "react/forbid-prop-types": 0,
