+++ package.json
@@ -7,8 +7,8 @@
     "build": "NODE_ENV=production webpack",
     "lint": "eslint \"./src/**/*.{js,jsx}\"",
     "start": "NODE_ENV=development webpack-dev-server",
-    "test-watch": "mocha ./_tests_/**/*.js --require babel-core/register --reporter nyan --watch",
-    "test": "mocha ./_tests_/**/*.js --require babel-core/register --reporter nyan"
+    "test-watch": "jest --watch",
+    "test": "jest"
   },
   "repository": {
     "type": "git",
@@ -16,6 +16,11 @@
   },
   "author": "Jon Oropeza <jon_oropeza@homedepot.com> (http://quotecenter.com)",
   "license": "Apache-2.0",
+  "jest": {
+    "setupFilesAfterEnv": [
+      "./jest/testSetup.js"
+    ]
+  },
   "bugs": {
     "url": "https://github.com/homedepot/react-component-grid/issues"
   },
@@ -26,6 +31,8 @@
     "@babel/plugin-transform-runtime": "7.4.4",
     "@babel/preset-env": "7.4.4",
     "@babel/preset-react": "7.0.0",
+    "@babel/register": "^7.4.4",
+    "@babel/runtime": "^7.4.4",
     "babel-eslint": "10.0.1",
     "babel-jest": "24.8.0",
     "babel-loader": "8.0.5",
@@ -33,17 +40,17 @@
     "babel-plugin-transform-es2015-spread": "6.22.0",
     "babel-plugin-transform-react-jsx": "6.24.1",
     "babel-polyfill": "6.26.0",
-    "chai": "^3.5.0",
-    "cheerio": "^0.20.0",
     "child-process-promise": "^2.0.1",
     "colors": "^1.1.2",
+    "enzyme": "^3.9.0",
+    "enzyme-adapter-react-16": "^1.13.0",
     "eslint": "^3.5.0",
     "eslint-config-airbnb": "^11.1.0",
     "eslint-plugin-import": "^1.9.2",
     "eslint-plugin-jsx-a11y": "^2.2.2",
     "eslint-plugin-react": "^6.3.0",
     "expect": "^1.20.1",
-    "mocha": "^2.5.3",
+    "jest": "^24.8.0",
     "react-hot-loader": "^1.3.0",
     "require-dir": "^0.3.0",
     "rimraf-promise": "^2.0.0",
